This commit introduces a simple Checkpointing API. It is the result of collapsing several intermediate revisions; the commit logs for those are below.
Collapsed revision
* merged.
* Merged dedalus-project/dedalus into default
* merged.
* Collapsed Ben's Checkpointing revisions:

* Moving checkpointing into new naming location.
* Functioning checkpoint restart; CFL calcs still broken and rescaling not tested.
* Implementing Daniel's suggestion for CFL calculation on restart.

As suggested in e-mail on 12/30/14.  Tested and functions but currently takes whatever is stored in self.stored_dt (initial_dt) as first step.  Will address in checkpointing.
* Working with Daniel; CFL now properly calculates correct CFL on restart.  Now just need to store prior dt in checkpoint.
* Adding memory to the RK timesteppers for the current dt.
* Working short-term solution to CFL.

Fixes lack of prior information about timestep size by avoiding doing min/max checks on first iteration.

Perplexing problem: first two timesteps seem to have same timestep.
* Daniel figured out source of repeated timestep on restart.  Also : minor output change (info->debug)

Issue is that we use the prior state to calculate the dt of the next step.  Normally, this is no issue, because we provide the intial step and then run from there.  On restart, this leads to two steps of same dt size, because first we evaulate the state to get a dt and then we use that same state (now stored as the prior step) to evaulauate and get a dt for step 2 (necessarily, same size as step 1).  This can be fixed by providing a dt on restart from a checkpoint file, or by doing our dt based on the current state.  Interesting.
* Merge.
* Merging to main branch and updates since spring.
* Merge.
* Adding rayleigh_benard test script for restarting from checkpoint.
* Cleaning up doc slightly.
* Main branch merge.
* Merged to main branch.
* Merging to main branch: bug fix for LBVPs and EVPs.
* Completed simple checkpointing API
* added find_output_counts; rewrote to use load_state.
* completed simple checkpointing API
* correcting errors in rebase process.
* added feature, suggested by Evan Anders, to have a list of directories
to exclude from checkpoint restarts. By default, checkpointing will
assume all subdirectories of the run directory except for
"checkpoints/" are Dedalus analysis directories to be restarted. The
excluded_dirs keyword to the Checkpoint object can contain a list of
directories to be ignored.
* fix by Ben to avoid bug in which max_writes was not respected after restart.
* merged.

hg hash: ee3f93a1ed41b85ecec5a48911d5e33a52e6847b